Convert things to a model where actions are kept in a hash separate from widgets. Change in implementation was surprisingly simple.
(MenuActions go to @actions, Theme::MenuActions still go to @widgets,
and @widgets_hash contains keys for both widgets and actions).

The exact details of this implementation need to be evaluated over
time.

The goal is to only keep real widgets in @widgets, to make things
a lot easier when implementing MVC-based widgets. (So that theoretically
the indices in @widgets would match indices in lists and/or container
widgets on the screen.)