Reworked prompter logic a bit to support actions that trigger as soon as e.g. Y or N is pressed, instead of having to press ENTER to confirm.
This is enabled by specifying 'instant: 1' among
ask() parameters.

As part of making this work, made hotkey: field
of MenuAction able to accept array of accepted
hotkeys (and these hotkeys can be gettext-ed!)

It works as-is now, but even better idea than
invoking the action as soon as any of the hotkeys
matches (and then doing sub-processing - based on
the specific key pressed - in the action handler)
would be to use a "<<"-similar idea to push
accepted keywords into prompter, e.g. like this:

prompt.text= 'Exit Taskman?'
prompt<< action( hotkey: N, function: (cancel))
prompt<< action( hotkey: Y, function: exit)

And then these possible choices would also be able
to appear in the menu below the prompt, like Pine
does it.

But, as said, works good enough as-is for now.