Support Find Next / Find Previous actions in the Index window
As part of the task, added function_arg, ability for a MenuAction to
specify parameters that will be passed to action when .run() is
invoked.

(Allows easier re-use of the same functions from multiple MenuActions,
e.g. it allows WhereIs, Find Next and Find Previous to be implemented
using the same function.)